Centralize errors/warnings/exceptions and update imports
