+++ src/ert/__main__.py
@@ -14,7 +14,7 @@ import yaml
 from ecl import set_abort_handler
 
 import ert.shared
-from ert._c_wrappers.config.config_parser import ConfigValidationError
+from ert._c_wrappers.config import ConfigValidationError
 from ert._c_wrappers.enkf import ErtConfig
 from ert.cli import (
     ENSEMBLE_EXPERIMENT_MODE,
