+++ src/ert/_c_wrappers/config/__init__.py
@@ -1,5 +1,6 @@
 from .config_content import ConfigContent, ContentItem, ContentNode
-from .config_parser import ConfigParser, ConfigValidationError, ConfigWarning
+from .config_errors import CombinedConfigError, ConfigValidationError, ConfigWarning
+from .config_parser import ConfigParser
 from .config_path_elm import ConfigPathElm
 from .content_type_enum import ContentTypeEnum
 from .schema_item import SchemaItem
@@ -16,4 +17,5 @@ __all__ = [
     "ConfigParser",
     "ConfigValidationError",
     "ConfigWarning",
+    "CombinedConfigError",
 ]
