+++ src/ert/_c_wrappers/enkf/_config_content_as_dict.py
@@ -1,8 +1,7 @@
 import logging
 from typing import Any, Dict, Optional, Tuple
 
-from ert._c_wrappers.config import ConfigContent
-from ert._c_wrappers.config.config_parser import ConfigValidationError
+from ert._c_wrappers.config import ConfigContent, ConfigValidationError
 from ert._c_wrappers.enkf.config_keys import ConfigKeys
 
 logger = logging.getLogger(__name__)
