+++ tests/unit_tests/c_wrappers/res/enkf/test_analysis_config.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ert._c_wrappers.config.config_parser import ConfigValidationError
+from ert._c_wrappers.config import ConfigValidationError
 from ert._c_wrappers.enkf import AnalysisConfig, ConfigKeys
